Exploring how to pull apart Excel files
This was a 30m spike.

One path is to have a binary on the user's computer that converts Excel
files. We can, in theory, rip apart the Excel files that exist (or a new
version thereof), and use them to generate an output format of our
choosing. This could be CSV, JSON, or other.

The binary could produce a file or files (.zip, .json) that the user
uploads to our service. We can wrap it as a drag-and-drop application,
where they take the Excel file, drag it onto our tool, and it either
says "good" or "you broke your Excel template." (We can do better than
that, obviously.)

The purpose here was to see what can be done with an existing form that
was submitted to the FAC. This code uses the single audit form
submitted by Berea College in 2021.

To run/test, you need a working Go environment, and probably need to

```go
go mod tidy
go run .
```

You can also pass an arbitrary file on input:

```go
go run main.go --filename <excel file>
```