Add UserProfile model
Establishing a Profile model to track FAC
specific data for a given user.

Our first use is as a temporary store of
form data leading up to the creation of an SF-SAC.
The Eligibility, Info, and Access responses are stored
here until they are used to create a new SAC instance.