First pass at content pages
This creates a cms app so that we can have a clear place for content pages. This will also help with assigning scoped permissions later.

There are 2 example pages to get the structure going. The home page lists the content pages. The content pages are at their own URL.

We would still like to move the base template out of the current app directory and in the template directory above. We can do that as we add the real content.