Add esbuild to docker build
Server & esbuild watch need to run in parallel

Point to compiled assets

Add entry point for JS

Configure postcss

Linting

Add flag to watch assets

`build-assets` defaults to not watching.
This is the version used in the Makefile.
Supplying the flag keeps esbuild running. This
is what's used in docker-compose