Adding environment variables
1) Everyone adds a local .env file for docker with a local setting - this is updated in the documentation
2) Add and env variable for GitHub actions
3) Add env variables to each cloud.gov space via manifest

The settings default to Prod, so that if something is configured on prod things like debug settings won't pop up. I still set it on production for explicitly for clarity.

This removes debug from our GitHub actions testing environment, once we do Pally Tests etc, this will be helpful so we don't anciently scan a debug page etc.

This should work in cloud.gov but I don't have the access to set up and test. We will need to verify the S3 bucket name in those environments.