Working dev deploy from GitHub Actions.
We moved npm run build to its own step in test_and_deploy.yml and put that ahead of the 18f/cg-deploy-action@main actions.
We switched back to using 18f/cg-deploy-action@main to push to cloud.gov and removed the debug step of trying to do that from a Docker container.
We put python manage.py collectstatic --noinput into the Procfile.
The result is that after having deleted all the assets in the bucket earlier, the last build (with the above changes) pushed the assets to the S3 bucket.
One of the key issues was that running collectstatic anywhere outside of the CF instance meant that Django couldn’t access the S3 credentials that are stored in VCAP_SERVICES, which is why we moved collectstatic back into the Procfile.