Run collect static twice
It is run by the buildpack and run in the Procfile.

Collectstatic needs to be run by the Procfile to get assets to s3. This adds images etc to the bucket, but not NPM built assets.
Collectstatic needs to be run by the buildpack so that the Procfile doesn't timeout.