Passing tests on new jsonnet FederalAwards
This commit... technically is committing compiled products into the
tree. However, I need FederalAwards.schema.json in the tree in order to
pass tests. So, one step at a time.

This commit introduces [jsonnet](https://jsonnet.org) into our tooling
pipeline. Jsonnet is an extended JSON that allows some useful things,
like comments, functions (pure, non-side-effecting), local variables,
and imports (well-scoped). These things allow us to build up the schema
as components that are composed as opposed to a single, monolithic file.

We will want to build those products as part of our bootstrap. This
could be done in Django using a Python library (which will build the
JSonnet files into .schema.jsons), or we could do it via a command-line
tool. The former will integrate into the stack, including letting us
test the build and then run unit tests against the resulting schema as
we go.

For iterative development of the schema, I typically did things like

pushd schemas
jsonnet FederalAwards.jsonnet > FederalAwards.schema.json
check-jsonschema --schemafile FederalAwards.schema.json test-files/federalawards-pass-01.json

This could/should be wrapped up into a Makefile for local dev and
testing.

This commit also reworks the test harness for this schema. The schema
has changed from its original design. In many cases, the validation for
the schema has been "pushed down." Now, instead of a massive chunk of
validations at the top level, the schema is made up of multiple objects,
each of which is responsible for its own validation. This makes it
easier to compose the schema as a whole, and makes it easier to reason
about the document/validations.

The excel mapping is *not yet fixed*. This means that a POST of an Excel
document will, currently, fail. Once the mappings are updated, however,
then uploaded documents should 1) conform to the schema and 2) pass (or
not) as apporpriate.