Spike on Excel generation
This adds code to:

1. Take a spreadsheet definition as a Jsonnet file
2. Render it (in memory) to JSON
3. Render it (to disk) as an Excel file

The `render.py` code should be turned into a Django command, so that it
can be invoked via `docker run web manage.py ...`

The test code is incomplete, but I am doing a commit for visibility. The
goal is to compare the named ranges in the schema and the spreadsheet,
and assert that they are equal sets. Or, otherwise, to guarantee that
the set of named ranges in the spreadsheet is neither larger nor smaller
than the set of named ranges in the schema.

This is work underway.