Breaking the FederalAwards schema
I think we need to rename some keys in the schema.

Why?

They're unclear. Although it was good to make the schema "deeper" (which
allowed for localization of the validations), changing things to "name"
based on object depth (thinking "this is part of the cluster object, so
name will obviously relate to the cluster") is not good. Because we're
now trying to validate/map back to the flat structure of a sheet, I need
more information in the field names.

The schema has changed ,which will break tests/validations in the
upload. We can catch that more systematically later. For now, I'm
aligning the sheet definition and the schema, and then we can fix the
mappings in excel.py.