Refactor env-egress module into a unified env module
The `env-egress` module was only invoking the `egress-proxy` module, and
it was making it harder to specify proxy values for use in the
`env-base` module without putting the burden on the including module.
So we just fold the invocation of https_proxy_instances = 1 the
`egress-proxy` module into `env-base`.

We also rename `env-base` back to `env` since there's no need to
distinguish it from neighboring modules anymore.