Exploring using schema to validate the API objects
PostgREST returns JSON objects by default.

We use JSON Schema throughout this project.

We could have schema for the objects that come back from the
disseminatino API. (In the case of 2023 data, we could actually use the
same schema that validate our input data.)

But, we don't have a way of checking the validity of what is coming back
from the API currently. This is a first step. I created an auditee
object, and can check that it contains well-formed data. Or, the
beginnings of this; more work would need to be done.

I can check for good zip code lengths/shape, as well as a decent UEI
check. (It's not perfect, but it is a good start.)

Gettnig this in the tree, so I don't lose it.