Adding a validate_api command
The changes to the Dockerfile/compose are twofold:

1. We need build-essential for jsonnet
2. We need to pass in an API key for api.data.gov

The requirements changes are for running Jsonnet. It is possible that
rqeuirements.in does not need to change, since we would not be running
these commands in production? Or, do we need it there, because we would
like to be able to run these tests in our GH Actions?

Everything else is to create a command that can be run as

fac validate-api --view vw_general --year 2022 --start 0 --end 100

and the command will hit the API and validate the objects against JSON
schema for those objects, reporting validation errors at the end.