Improved schema, patched data
To experiment with improved regex in the schema (for email, phone), I
patched the data as it came back.

We should correct the data inbound. We need to decide if it is OK to
format phone numbers to conform with E164. This would be a good time to
do it, and standardize the FAC on... a standard.