Update actions/checkout to v3
[node 12 is deprecated, and actions that were using it have newer versions out.](https://github.blog/changelog/2022-09-22-github-actions-all-actions-will-begin-running-on-node16-instead-of-node12/)

We're still using older versions of those actions and as a result we're seeing messages like the following in our GitHub Action runs:
```
test-and-lint / test
Node.js 12 actions are deprecated. Please update the following actions to use Node.js 16: actions/checkout@v2, actions/setup-node@v1. For more information see: https://github.blog/changelog/2022-09-22-github-actions-all-actions-will-begin-running-on-node16-instead-of-node12/.
```

This changes upgrades the versions of the actions we're using to ones that use node 16.