Add support for VSCode debugging
This does a few things:

1. Fixes some settings regarding minio and clamav so that scanning and
   saving of files works in the local stack. (This may conflict with
another branch/merge...)
2. Adds `tblib` for traceback support in the container.
3. Modifies run.sh so that we can run (optionally) with debugging
   support enabled; set VSCODE_DEBUG in .env (or your environment)
before your compose to enable.

https://testdriven.io/blog/django-debugging-vs-code/
https://www.yellowduck.be/posts/debugging-django-vs-code
https://code.visualstudio.com/docs/python/tutorial-django

I'm sad to say I did the debugger config quickly; we'll want to document
it for others. `launch.json` gets created by VSCode, and that lets you
connect the debugger to the containerized stack on 5678. Note that when
running with the debugger, Django does not start until you start the
debugger... in other words, the stack comes up, but the web stack sits
there waiting for you to start the debugger in VSCode before kicking off
the Django stack.

Pushes against this branch to improve the commit are welcome. This was
fast work early in the morning to unblock development.