Add initial help messages
This commit adds text to the help messages. We will need to revisit two
things:

1. The content of these help messages.
2. All fields where `Help.unknown` was used.

In the second case, I did not know what help message should be used in
all cases. Therefore, we will need to do a walk through the messages
(and how they are attached) before we test with users.