Paying down tech debt
This breaks out the parsing of the Jsonnet specs from processing them.
The result of parsing is a collection of NamedTuples, which gives us a
layer of ... protection. If we can't build a good named tuple, things
fail.

Once we have the NTs, we then walk them to produce the workbook. The
changes there in this commit are entirely to use the NTs instead of
trying to reach into a JSON object. So, no functional changes at this
point.

`parse.py` is introduced to keep the parsing/NT work separate

A later iteration should turn this into a Django command.