Updated build, fixed parse/render
1. The top-level Makefile now does everything front-to-back.
2. Added `merged_unreachable` as a NT, which was not handled before.

For 1, this now regenerates JSON source documents (program names, ALNs),
which come from the DB. Eventually, we will likely pull from our DB via
API to generate these lists. (Or, something.) For now, we procress data
from Census extracted from their dumps.

For 2, this wasn't handled becuase the sheet I was working with did not
have this field/type. I added it, parsed it, and incorporated it into
the script. (`merged_unreachable` was already in `render.py`, but
`parse` was not handling it correctly. Given that `parse.py` was new,
this makes sense to me.)