Fixes FederalAwards schema, tests
There were tests that were incorrect based on my reading of the
documentation in the Census sheet.

This corrects the schema to reflect that reading, as well as corrects
the tests.

There is a question whether I made anything *too permissive*, and more
testing now will be necessary.

This allows the workbook that CS developed for testing to pass. "Allows"
suggests I permitted something; moreso, "the workbook that was developed
now appears to correctly passed, based on the documentation available to
me."