Rudimentary data processing for Census dumps
This is bad. But, it gets us started.

This takes a set of CSVs from Census, processes them, loads them
locally, and dumps them.

1. The CSVs can be found in GDrive.
2. The _tables.sql document converts the Oracle schema to a Postgres
   schema in the rudest way possible.
3. The processing strips repeated headers, and makes sense of multi-line
   cells. This is not rigorously validated (or, at this point, *at
all*).
4. The data is loaded via \copy, which is the fastest way to load data
   from a CSV into tables in Postgres. Note that this is fragile based
on the order of the names in the schema and CSV headers.
5. The tables are dumped for movement to cloud.gov, if we want.

For now, this provides a fast path to loading this data locally for
exploration.