Jadudm/additional xlsx in book validations (#1345)
* Paying down tech debt

This breaks out the parsing of the Jsonnet specs from processing them.
The result of parsing is a collection of NamedTuples, which gives us a
layer of ... protection. If we can't build a good named tuple, things
fail.

Once we have the NTs, we then walk them to produce the workbook. The
changes there in this commit are entirely to use the NTs instead of
trying to reach into a JSON object. So, no functional changes at this
point.

`parse.py` is introduced to keep the parsing/NT work separate

A later iteration should turn this into a Django command.

* Initial commit of new validation for workbooks

This adds the ability to load columns of data into sheets and then
validate against those columns in dropdowns. This lets us have lists of
up to 65K elements for validation. (Or, perhaps it was 32K.)

Either way, we can now validate all ALNs and program names.

* Updated build, fixed parse/render

1. The top-level Makefile now does everything front-to-back.
2. Added `merged_unreachable` as a NT, which was not handled before.

For 1, this now regenerates JSON source documents (program names, ALNs),
which come from the DB. Eventually, we will likely pull from our DB via
API to generate these lists. (Or, something.) For now, we procress data
from Census extracted from their dumps.

For 2, this wasn't handled becuase the sheet I was working with did not
have this field/type. I added it, parsed it, and incorporated it into
the script. (`merged_unreachable` was already in `render.py`, but
`parse` was not handling it correctly. Given that `parse.py` was new,
this makes sense to me.)

* Moved requirements.txt

Should have been part of the previous commit.

* Ignore random XLSX documents in this tree

Hopefully this does not ignore down the tree? Perhaps it does, in which
case, this might want to be removed.

* Merging main... into main...

* Paying down tech debt

This breaks out the parsing of the Jsonnet specs from processing them.
The result of parsing is a collection of NamedTuples, which gives us a
layer of ... protection. If we can't build a good named tuple, things
fail.

Once we have the NTs, we then walk them to produce the workbook. The
changes there in this commit are entirely to use the NTs instead of
trying to reach into a JSON object. So, no functional changes at this
point.

`parse.py` is introduced to keep the parsing/NT work separate

A later iteration should turn this into a Django command.

* Merging main... into main...

* Updated workbooks. No other changes.

* Update federal-awards-expended-template.jsonnet

I was using the `yorn` help text for the cluster name column... this is incorrect.

* Fixed linting and complexity issue

* Regenerated output files

* Removed unused file

* testing linting fix on mypy

---------

Co-authored-by: Hassan D. M. Sambo <hassandeme.mamasambo@gsa.gov>
Co-authored-by: Alex Steel <alexander.steel@gsa.gov>