Fixing headers in schema
We don't know that we will receive CSVs with headers in the same order
every time. If not, we need to mangle our schema to match the order so
that we can use `\copy`.

Ultimately, this all wants to be production-ready ETL code, not my
hacked scripts. This is for proof-that-we-can-use-the-data only, not
production.