Moving the new API into v1.0.0-beta (#1482)
* Moving the new API into v1.0.0-beta

This creates a path to multiple APIs active at once, and versioning
them.

* Fixing linting.

* Fixing tests

* Linting, fixed views.

Tests pass.

* Fixing test url.

* The URL is broken.

Needed for demo.

* add api version everywhere

* remove prints

* Trying to get tests to pass live

I have no idea why these migrations appeared.

I really, really don't like this branch/code.

* Adding back in the other test.

* Remembering the import...

Honestly.

* Splitting things out

This splits things out, and makes it possible to reuse the list of live
and deprecated APIs, as well as reuse some of the construction functions
for testing.

Pushing for discussion.

* Separates schema and view creation

This was requested by JMM to work around migrations. We need the schemas
before we need the views.

* Forgot to deprecate schemas in .profile

* Fixing type hinting.

I need to learn how to run the lint tests on a single directory. As
configured, they run against my entire tree, including things that are
not our code.

---------

Co-authored-by: danswick <daniel.swick@gsa.gov>