Moving towards JWT encoded/protected API.
This breaks out a base, where we create users. Currently, we are riding
on the fact that 'anon' and friends were created long ago. They are not
part of our setup/teardown process.

There are some questions around how we setup/teardown users and views.
This works, but it isn't a full-up/full-down yet.

A command to create a signed JWT exists, but we need to do TF work and
GH Secret work to actually integrate it into the stack. Not there yet.