Attempting to run data through a test process
This attempts to load workbooks generated from the generator script
through the app itself. The goal would be to load all of the workbooks
in a set into the app, trigger dissemination/ETL, and then validate data
is in the mocked DB.

This is a first commit of the command; it might become a test instead,
or it might remain a command. For now, just trying to explore what could
be.