Runs through five workbooks.
This is finding possible errors in validations and/or workbook
generation.

That can be another ticket/branch. However, for now, what is good is
that this is letting me run workbooks through Django, see them end up in
minio, and they go throguh the validation process.

Next will be to see what is required to trigger dissemination.