Working to validate 100010
The cluster name validation logic was broken.

There may still be more issues with it, but this makes it clearer, I
think, and fixes it.

The issue has to do with `0` always coming in for `cluster_total`. That
means we have to either 1) always have a cluster_name field present, or
handle it in the validation.