Adding a PDR about versioning
We need documentation about how we version, and to record that we will
make changes to workbooks that (sometimes) will be very inconvenient for
users.

This lays out an initial process for discussion. If accepted, we will
move the documentation to the web. This will suggest further editing and
plain language work.

We will update this PDR with improved, plain language text (or issue a
new PDR) to reflect the final version that we work with.