Fixing the cluster name validation
These changes fix our cluster name validation.

1. A valid cluster name is either:
    1. N/A
    2. One of the cluster names
    3. STATE CLUSTER
    4. OTHER CLUSTER NOT DEFINED
2. IF a user selects N/A or a cluster name, then both the state cluster
   and other cluster fields MUST be empty.
3. IF a user selects STATE CLUSTER, then the state cluster name must be
   completed, and the OTHER field must be empty.
4. IF a user selects OTHER CLUSTER, then the state cluster name must be
   empty, and the OTHER field must be completed.

Our previous validation was more confusing, and we may not have had all
of the choices present.

This means we can use a rigid validation in our workbooks for state
cluster. The user has the ability to pick OTHER and enter data in the
appropriate fields. We should NOT allow them to edit the cluster name
field in our workbooks.

This validation should also prevent submission if they do, or manage to.