Adding tools for workbook generation and testing
This brings two tools into the tree that have been used for exploring
workbook generation.

`generate-sqlite-files` takes public, pipe-delimited Census data and
turns it into an SQLite3 database. This improves on previous tools that
did this in the past.

`workbook-generator` is a script that takes a DBKEY and an SQLite3
database containing public Census data, and outputs a set of GFAC-style
XLSX workbooks containing data.

To do this, `workbook-generator`

1. Loads one of our templates using `openpyxl`
2. Loads data from the SQLite3 database into the named ranges in the
workbook template
3. Saves the template out to a new filename, populated with data.

The generator attempts to do everything right. For example, it attempts
to follow old linking IDs (e.g. ELECAUDITID) and replace them with
AWARD-####-style references. Similarly, it unpacks the odd design of the
Notes to SEFA tables to convert the Census data back into a functional
workbook.

This is not a fully validated tool. However, it does generate workbooks
with authentic data, and they have been used to drive upload processes
into our system. Using them, we have found errors in our validations
that we did not know about previously.

If we continue using this tool, we will likely want to build it into our
testing automation process. In theory, we could have hundreds (or
thousands) of workbooks generated and ready for testing. We can also
generate workbooks that explicitly test (e.g.) having secondary auditors
(or not), or that exhibit specific other properties... all from
existing, previously-validated data.

The generator also spits out a JSON document. That document records:

1. The table data was pulled from
2. The fields pulled
3. The values in those fields

The purpose of that document is to be able to do something like:

1. Load the workbooks through our pipeline
2. Have them undergo validation, cross-val, and ETL
3. Use the JSON document to compare what we pulled from the SQLite DB to
what ended up in our dissemination DB.

In other words, the JSON document is to enable end-to-end testing of the
dissemination pipeline. Ideally, we would do that final check using the
API. This would let us use the JSON document to generate API calls that
query the DB (from the "outside"), and verify that the API produces data
we expect.