Jadudm/pdf to model (#1736)
* Adds PDF to model

Includes the migration, validations, and tests. Adds in the component
page numbers from the form.

* Linting.

* Linting.

* Fixing linting.

Had to add a type to force some testing that was intended to fail.

* Add in PGRST secret to test env

* Backing out changes.

This branch had changes that were unrelated to the feature under
development. This backs them out to sync with `main`.

* Another deviation from main

This was a brief attempt to have multiple test files against the schema.

* Also reverting from main

* Update backend/schemas/source/base/Base.libsonnet

* Update backend/schemas/source/base/Base.libsonnet

* Update backend/schemas/source/base/Base.libsonnet

* Removing unneeded ref to audit_report

The PDFs are stored in a different part of the model. Removing
commented-out confusion.

* Removing for circular ref

With this here, we end up with a circular ref frmo within the checklist
code.

* Undo.

This circular ref comes in as part of a different branch, and the change
should not be present here.

* Making sure we don't break on an absent report

Added a try/catch to the search for the audit report.

* flake8 to the rescue