Disable schedule for releases
Per discussion with Jadud, before we automate the release process,
we want to ensure that regression tests are more verbose, and have a
recovery process. If a migration occurs that blows up the data, we do
not want to promote that migration. We want to automate a way to
migrate the data and sync the prod db back into staging for disaster
recovery scenarios.