Adding in "safe-by-design" intake->dissem
This did a few things.

1. I made the name more descriptive. We can change it back.
   I did this in part because we will have other kinds of ETL in the
future, hence making this one I->D.
2. The class now loads objects and does not save them by default. It
   holds them internally.
3. Added explicit methods to `get` those stored objects, as well as
   `save` those stored objects.

With minor modifications, all tests passed once the code was changed
from a single `load` call to a `load` followed by a `save`.

I also have a test that loads and confirms nothing touches the database.

These changes were based on feedback from code review. The decision was
to make the default behavior that we do *not* save data. Why? Because we
want to use this in other parts of the codebase (for
flattening/enmodelnig the data that is in the intake DB), and we
absolutely do not want to accidentally trigger the save. This makes it
so that the save action is very explicit.