Enhance Staging Automation (#1757)
* This adds a repository event for usage with autodeploy

* Set up different tokens to author and approve the PR

Switched GITHUB_TOKEN to approver and use a DEPLOY_TOKEN to author

* Format `gh pr create` and add a new label

"automerge" label added to ensure action reads it.
A new custom label needs to be added prior to running

* Setup automerge to run when a PR meets a certain criteria

* Create an automated release

GITHUB_TOKEN may need to change to DEPLOY_TOKEN
Will run on Wed at 2p EST and checkout prod branch

* Add dependency on testing

* Disable schedule for releases

Per discussion with Jadud, before we automate the release process,
we want to ensure that regression tests are more verbose, and have a
recovery process. If a migration occurs that blows up the data, we do
not want to promote that migration. We want to automate a way to
migrate the data and sync the prod db back into staging for disaster
recovery scenarios.