Make event targets consistent across event types
The DOM manipulation code was fragile and complicated
because it was trying (and failing) to account for the
node that was the event target and then traverse back
up. Just making the element with the event listener on
it be the one we base the DOM traversal on makes it all
simpler and less fragile (I hope?).