Make data loading transparent and repeatable (#2079)
This brings the container back that loads historic data for dev and
test.

Ultimately, this all has to be replaced, becuase the data will change
over the course of the next three months.

The current load solution does not work well locally for two reasons:

1. It loads the container stack in the background, making it hard to
   debug via the container logs
2. It is not repeatable, because `load` assumes that it will only ever
   run once.

This brings back the container that loads the data. It works as part of
the stack, and was designed to work repeatably. While it is *different*
from how the GH action runner behaves, this works every time in the
local stack.

We will treat aligning the deploy and local environments as tech debt.
This is blocking devs from doing their work. This unblocks devs.