2062 add additional ein to dissemination (#2084)
* An ETL class that returns values.

* Fixed tests.

* Adding a method to get the loaded objects.

* Adding in "safe-by-design" intake->dissem

This did a few things.

1. I made the name more descriptive. We can change it back.
   I did this in part because we will have other kinds of ETL in the
future, hence making this one I->D.
2. The class now loads objects and does not save them by default. It
   holds them internally.
3. Added explicit methods to `get` those stored objects, as well as
   `save` those stored objects.

With minor modifications, all tests passed once the code was changed
from a single `load` call to a `load` followed by a `save`.

I also have a test that loads and confirms nothing touches the database.

These changes were based on feedback from code review. The decision was
to make the default behavior that we do *not* save data. Why? Because we
want to use this in other parts of the codebase (for
flattening/enmodelnig the data that is in the intake DB), and we
absolutely do not want to accidentally trigger the save. This makes it
so that the save action is very explicit.

* Added latest changes from main

* regenerated output

* Work in progress

* Work in progress

* FederalAwards is completely revised.

* Removed `Revisions`

This table is not in use currently. So, it should not be here (yet).

* Notes to SEFA done.

1. Checked fields.
2. Modified the minimis field to take Y/N/Both in the validation.
3. Updated base to contain this new Enum.
4. Updated help text and validations in the workbook; checked.
5. Updated the workbook to use this new defn for the field.
6. Updated model to use text fields.
7. Updated Intake2Dissem to use 'get' and pass an empty list if no
   entries.
8. Added a requirement on the 'rate_explained', because... it seems to
   always be required.

Looks complete.

* walk captext

* walk secondaryauditors

* possibly enforce text length for state abbrevs

* Removed `Revision` from the model, moved code around.

No major changes yet. Committing to remain in sync with others. WIP.

* Walked `Findings`.

Tests need to be run on the schema changes.

* Updating for WIP w/ HDMS

* Updated general dissemmination etl

* Updated to match dissemination

* Refactored general dissemination model

* Updates to pass some tests with new schema

* Updated output files

* Fix linting issue

* More linting

* Getting tests to pass after updating schema validation rules

* Regenerated output

* Still cannot get all the excel tests to pass

I'm unclear what we're triggering in the schema..

* Passing tests, but...

I want us to look more at the validation for `is_guaranteed.`

* Fixed validation on loan guarantee

Put the OneOf in the properties.

This simplifies the case wehre there must be a value. I could not get it
to validate otherwise. Now it looks for an N/A, or a string of numbers.

* Using the constant for N/A

* Adding test fixture updates

* Made GeneralInformation Schema more Permissive

* Updated workbook templates

* Updated certification date schema and made certification parent schema more permissive

* Fixed validation rule for loan_balance_at_audit_period_end

* Fixed validations/schemas/test cases

* Linting

* Migration for cognizant fields

* Adding in views for API, historic API

The API needs to sync with the model, so it makes sense to bring it in
here.

The historic API might as well come in, too, because we're bringing the
data into `main` elsewhere. Once the data exists, the API can/should
exist.

* Fixes

1. Fixes the API views. The names were wrong in places.
2. Fixes the order of function and view creation. The views rely on the
   functions existing.
3. Adds a more verbose run sequence to `run.sh`, which is lifted from
   `.profile`. Really, we should break that out someday.

* Rename v1_0_0_beta to v1_0_0

* Pulled in changes from main

* Linting

* Code cleaning

* Reverted certification schemas back to fully required mode

* Cleaning

* Updated schema to ensure loan balance still validates properly when this value comes in as text

* Reset migrations

* Linting

* Testing a fix for pa11yci issue

* Revert "Testing a fix for pa11yci issue"

This reverts commit 092b92137f3df223c4bf368a1d07f412d7551bf9.

* Regenerated workbook

* Updated schema to allow empty string for auditor_state when non-USA address is selected

* Remove null condition from dissemination model

* #2062 Added dissemination logic for additional eins

* #2062 Updated test cases

* #2062 Added Django migration

* Linting

* #2062 Switched from single save to bulk save

* #2062 Removed unnecessary test cases

---------

Co-authored-by: Matt Jadud <matthew.jadud@gsa.gov>
Co-authored-by: danswick <daniel.swick@gsa.gov>