Backup Database and put in s3 (#2109)
* Add backup cf task

Should only run if a tag is present

* Adds django backup

* adds django backup to requirements

* fix DBBACKUP_STORAGE

* Set to private bucket

* Fix linting issue

* move DBBACKUP

* move backup into private s3 chunk

* remove db backup?

* add dbbackup back

* move into other INSTALLED_APPS

* Fixes settings.py

* pg_dump needs to match server/client versions

pg_dump: error: server version: 15.3; pg_dump version: 14.9 (Ubuntu 14.9-0ubuntu0.22.04.1)
pg_dump: error: aborting because of server version mismatch

* Testing if data load works with psql 15

* both psql client versions?

* try postgresql-common

* remove pg_common

* Bump to jammy psql v15

* Add comments to apt.yml

* Remove unnecessary path

* add callable script for cf tasks

* Update to script and make +x

* call script instead of direct command

* Make db dumps a bit more readable

* add correct url to apt.yml comment

* rollback

* Bump Mem to 1G

In order to run python manage.py dbrestore, 512M wasn't enough to actually
do the restore.

* Set a readable name to the db dumps (again)

* Set back to 1G like the other envs

* Set app instance back to 512M Mem

* Sketch: Add a bucket for backups, shared to appropriate spaces

NOTE: NOT YET FUNCTIONAL! This is a sketch until the TODOs are addressed

* Modify settings for new bucket

* Bind the bucket to the running app

This makes it so that we do not need to modify the manifest, and only
expose the backup bucket while we are actually using it, i.e when we want to
do a backup.

* terraform fmt

* Shares the s3 backup bucket with designated spaces

Based on the values in config.tf

* Fix typo on unbind-service cf command

* Regenerate TF Plan

* Regenerate TF Plan

* Remove django-storages

This appears to be redundant. Removing this does not change the
requirements.txt file

* Add default behavior for dbbackup if local env

* Fix linting issue

---------

Co-authored-by: Bret Mogilefsky <bret.mogilefsky@gsa.gov>