Only allow certain users to run restore workflow
This can be a dangerous operation, and we want to restrict who can run it.
Since there isn't a way to limit it to a group (fac-admins), we are specifying specfic
users to run this workflow. As of right now, asteel/JMM are the only ones who can run this, allowing us to test
the operation in dev. Once this is set, the list will contain all users from fac-admins.

The codeowners has also been updated to watch this specific file, so that
fac-admins and asteel are aware if this file is changing.