Jmm/aws s3 alternative (#2296)
* Restore Database Script

* Make script executable

* Preliminary backup and restore workflows

* disallow backups in staging

* Only allow backups in dev

This workflow is exclusively for backing up development environment.

* Remove full name, and only supply date

After discussing with JMM, this seems like a good safeguard to prevent against
misaligned filenames, and build the file name with the inputs

* bump to 3g disk, just to be safe for the tmp files

* Restore Database Script

* Make script executable

* Preliminary backup and restore workflows

* disallow backups in staging

* Only allow backups in dev

This workflow is exclusively for backing up development environment.

* Remove full name, and only supply date

After discussing with JMM, this seems like a good safeguard to prevent against
misaligned filenames, and build the file name with the inputs

* bump to 3g disk, just to be safe for the tmp files

* Allow dev and prod to be backed up via workflow

* Only allow certain users to run restore workflow

This can be a dangerous operation, and we want to restrict who can run it.
Since there isn't a way to limit it to a group (fac-admins), we are specifying specfic
users to run this workflow. As of right now, asteel/JMM are the only ones who can run this, allowing us to test
the operation in dev. Once this is set, the list will contain all users from fac-admins.

The codeowners has also been updated to watch this specific file, so that
fac-admins and asteel are aware if this file is changing.

* First version of fac_s3

* Merge and Lint

* Test fac_s3 locally with dbbackup

---------

Co-authored-by: Alex Steel <130377221+asteel-gsa@users.noreply.github.com>
Co-authored-by: JeanMarie Mariadassou <jmpm@JMPMs-MBP.local>